Broke down GameState into BrickBreakerGameState and GalagaGameState
Still need to change some things but the structure is there
