+++ src/GameState.java
@@ -7,45 +7,24 @@ import java.util.Random;
 import javafx.scene.Group;
 import javafx.scene.Node;
 
+//Edited by Ethan Jeffries from GameState class in BrickBreakers
+//Changed functionality to be a super class which will have children classes for each arcade game
 public abstract class GameState {
 	
-	
-	private final int brickWidth = 50;
-	private final int brickHeight = 15;
-	private final int maxBrickOdds = 100;
-	private final int maxUnbreakableBrickOdds = 100;
-	private final int bricksPerRow = 8;
-	public Paddle paddle;
-	public List<Brick> bricks;
-	public List<Ball> balls;
 	public List<GameElement> gameElements;
 	private int screenWidth;
 	private int screenHeight;
 	private Group root;
-	private List<PowerUp> powerUps;
 	private LifeCounter livesLeft;
 	private ScoreCard score;
 	private boolean gameLost;
-	private final int oddsToGeneratePowerUp = 1001;
-	private int currentStep = 0;
-	
 	
 	public GameState(int screenWidth, int screenHeight) {
 		this.screenWidth = screenWidth;
 		this.screenHeight = screenHeight;
 		
 		this.root = new Group();
-		
-		balls = new ArrayList<Ball>();
-		bricks = new ArrayList<Brick>();
-		powerUps = new ArrayList<PowerUp>();
-
-		this.paddle = new Paddle(screenWidth, screenHeight);
-		root.getChildren().add(this.paddle.getNode());
-		
-		//Changed the ball input here so it would be in the middle of the screen
-		this.addBall();
-		
+		spawnGameItems();
 		
 		this.score = new ScoreCard(screenWidth, screenHeight);
 		root.getChildren().add(score.getNode());
@@ -57,8 +36,10 @@ public abstract class GameState {
 		
 	}
 	
-	//created class to construct bricks 
-	//takes in values for number of rows and bricks per row as well as the odds of a brick which will be unique to each level
+	//Empty method that is inherited and overridden by children classes for creating game items
+	public abstract void spawnGameItems
+	
+	//This method WILL SOON BE CHANGED TO SPAWN TARGET CHARACTERS REGARDLESS OF GAMESTATE
 	public void makeBricks(int numberOfBrickRows, int oddsOfBrick, int oddsOfUnbreakableBrick) {
 		//creating the bricks 
 		Random randomVal = new Random();
@@ -80,6 +61,7 @@ public abstract class GameState {
 		}
 	}
 	
+	//THESE TWO MOVEMENT FUNCTIONS WILL BE CHANGED TO BE GENERAL FOR THE MOVER CLASS
 	public void moveLeft() {
 		this.paddle.moveLeft();
 	}
@@ -88,106 +70,19 @@ public abstract class GameState {
 		this.paddle.moveRight();
 	}
 	
+	//Changing to be more general for gamestates
 	public void step(double elapsedTime) {
-		if (++ currentStep % oddsToGeneratePowerUp == 1) this.generatePowerUp();
-		for (int i = 0; i < balls.size(); i++) {
-			balls.get(i).move(elapsedTime);
-			balls.get(i).bounceOffWall(screenWidth, screenHeight);
-		}
-		this.bounceOffBricks();
-		this.bounceOffPaddle();
-		this.checkForPowerUps();
-		this.checkBallIsOut();
-		this.checkForBalls();
+		makeGameChecks();
 	}
+	
+	//Method that is inherited and overridden in child classes to make frame checks for objects
+	public abstract void makeGameChecks
 	
 	public Group getRoot() {
 		return this.root;
 	}
 	
-	public void bounceOffBricks() {
-		for(Ball currentBall: balls) {
-			for(int i = 0; i < bricks.size(); i ++) {
-				Brick currentBrick = bricks.get(i);
-				boolean intersects = currentBall.getBounds().intersects(currentBrick.getBounds());
-				if(intersects) {
-					//Change logic so it checks the position of the ball relative to the brick and bounces if the middle is outside of the brick
-//					if (currentBall.getBounds().getCenterX()>currentBrick.getBounds().getMinX() && currentBall.getBounds().getCenterX()<currentBrick.getBounds().getMaxX()) {
-//						currentBall.bounce();
-//					}
-//					else {
-//						currentBall.sideBounce();
-//					}
-					currentBall.bounce(currentBrick);
-					if(currentBrick.getBrickType().equals("brick")) {
-						System.out.println("hit and removed a brick");
-						score.incrementScore(100);
-						this.root.getChildren().remove(currentBrick.getNode());
-						this.bricks.remove(currentBrick);
-					}
-					else {
-						System.out.println("unbreakable");
-					}
-					
-				}
-			}
-		}
-	}
-	
-
-	public void bounceOffPaddle() {
-		for (Ball currentBall : balls) {
-			currentBall.bouncePaddle(this.paddle);
-		}
-	}
-
-	public void checkForPowerUps() {
-		for(int j = 0; j < balls.size(); ++j) {
-			Ball currentBall = balls.get(j);
-			for(int i = 0; i < powerUps.size(); i ++) {
-				PowerUp currentPowerUp = powerUps.get(i);
-				boolean intersects = currentBall.getBounds().intersects(currentPowerUp.getBounds());
-				if(intersects) {
-					System.out.println("HIT POWERUP");
-					this.root.getChildren().remove(currentPowerUp.getNode());
-					currentPowerUp.activatePowerUp();
-					this.powerUps.remove(currentPowerUp);
-				}
-			}
-		}
-	}
-	
-	public void checkBallIsOut() {
-		ArrayList<Ball> badBalls = new ArrayList<Ball>();
-		for (int i = 0; i < balls.size(); i ++) {
-			Ball currentBall = balls.get(i);
-			double ballY = currentBall.getBounds().getCenterY();
-			if (ballY>screenHeight){
-				badBalls.add(currentBall);
-				balls.remove(i);
-				this.root.getChildren().remove(currentBall.getNode());
-			}
-		}
-	}
-	
-	public void checkForBalls() {
-		if (balls.size() == 0 && this.livesLeft.getLivesLeft() > 0) {
-			this.livesLeft.changeLives(-1);
-			if (this.livesLeft.getLivesLeft() > 0) {
-				System.out.println(String.format("You have %d lives left, good luck.", this.livesLeft.getLivesLeft()));
-				this.addBall();
-			}
-		} else if (balls.size() == 0) {
-			gameLost = true;
-			this.root.getChildren().add(new GameOverMessage(screenWidth, screenHeight, this.score.getCurrentScore(),
-					this.score.getHightScore()).getNode());
-		}
-	}
-
-	public void increasePaddleSize() {
-		this.paddle.increaseX();
-	}
-	
+	//Method checks for game over
 	public boolean gameLost() {
 		return this.gameLost;
 	}
@@ -197,28 +92,12 @@ public abstract class GameState {
 		return (bricks.size()==0);
 	}
 	
-	public void addBall() {
-		Ball newBall = new Ball(screenWidth, screenHeight);
-		balls.add(newBall);
-		root.getChildren().add(newBall.getNode());
-	}
-	
+
+	//Will soon be changed to if the number of targets = 0
 	public boolean isWon() {
 		return this.bricks.size() == 0;
 	}
 	
 	
-	public void generatePowerUp() {
-		int rand = this.paddle.getRandomInRange(0, 100);
-		if (rand < 15) {
-			AddBallPowerUp powerUpToAdd = new AddBallPowerUp(screenWidth, screenHeight, this);
-			root.getChildren().add(powerUpToAdd.getNode());
-			powerUps.add(powerUpToAdd);
-		} else if (rand < 30) {
-			IncreasePaddleLength powerUpToAdd = new IncreasePaddleLength(screenWidth, screenHeight, this);
-			root.getChildren().add(powerUpToAdd.getNode());
-			powerUps.add(powerUpToAdd);
-		}
-	}
 	
 }
